pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = [
    {name = "Egor Titov",email = "titov.egor1001@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask (>=3.1.1,<4.0.0)",
    "flake8 (>=7.2.0,<8.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "validators (>=0.35.0,<0.36.0)",
    "requests (>=2.32.4,<3.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
]

[tool.poetry]
packages = [
    { include = "page_analyzer" },
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
